requirements.txt
boto==2.48.0
boto3==1.4.7
botocore==1.7.33
certifi==2017.11.5
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11.4
django-allauth==0.34.0
django-appconf==1.0.2
django-braces==1.12.0
django-comments-xtd==2.0.9
django-contrib-comments==1.8.0
django-crispy-forms==1.6.1
django-dropbox==0.0.2
django-dropbox-redux==0.0.5
django-dropbox-upload-handler==0.1.1
django-imagekit==4.0.2
django-model-utils==3.0.0
django-star-ratings==0.6.0
django-storages==1.6.5
djangorestframework==3.6.4
docutils==0.14
dropbox==8.2.0
gunicorn==19.7.1
html5lib==0.9999999
idna==2.6
jmespath==0.9.3
Markdown==2.6.9
oauthlib==2.0.6
olefile==0.44
pilkit==2.0
Pillow==4.2.1
psycopg2==2.7.3
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
s3transfer==0.1.12
six==1.11.0
swapper==1.1.0
urllib3==1.22
whitenoise==3.3.1
